[DRT-5078] - Desk deployments for tomorrow are showing zero
